+++ python/paddle/distributed/cloud_utils.py
@@ -14,7 +14,7 @@
 
 import os
 import paddle
-from paddle.distributed.utils import get_cluster, logger
+from paddle.distributed.utils import get_cluster, logger, get_gpus, get_cluster_from_args
 
 
 def get_cloud_cluster(args_node_ips, args_node_ip, args_port, selected_gpus):
@@ -92,3 +92,27 @@ paddlecloud environment.".format(args_node_ips, node_ips))
     cluster, pod = get_cluster(node_ips, node_ip, trainer_endpoints,
                                selected_gpus)
     return cluster, cluster.pods[node_rank]
+
+def _get_trainers_num():
+    return int(os.getenv("PADDLE_TRAINERS_NUM", "1"))
+
+def get_cluster_and_pod(args):
+    # parse arguments, used for cloud-single-machine and local
+    selected_gpus = get_gpus(args.selected_gpus)
+    trainers_num = _get_trainers_num()
+    logger.debug("parsed from args trainerss_num:{} selected_gpus:{}".format(
+        trainers_num, selected_gpus))
+
+    cluster = None
+    pod = None
+
+    if args.use_paddlecloud and trainers_num != 1:
+        cluster, pod = get_cloud_cluster(
+            args.cluster_node_ips, args.node_ip, args.started_port,
+            selected_gpus)
+        logger.info("get cluster from cloud:{}".format(cluster))
+    else:
+        cluster, pod = get_cluster_from_args(args, selected_gpus)
+        logger.info("get cluster from args:{}".format(cluster))
+
+    return cluster, pod
